Distinct AST concepts
