--- src/ast.ts
@@ -1,164 +1,0 @@
-import { parse } from "@babel/parser";
-import traverse, { TraverseOptions, NodePath } from "@babel/traverse";
-import * as t from "@babel/types";
-import * as recast from "recast";
-
-import { Code } from "./editor/i-write-code";
-
-export { NodePath } from "@babel/traverse";
-export * from "@babel/types";
-export { traverseAST, transform, Transformed };
-export {
-  ASTSelection,
-  ASTPosition,
-  isSelectableNode,
-  isSelectableVariableDeclarator,
-  isSelectableIdentifier,
-  SelectablePath,
-  SelectableNode,
-  SelectableObjectProperty,
-  SelectableIdentifier,
-  SelectableVariableDeclarator,
-  Selectable
-};
-export {
-  isArrayExpressionElement,
-  areAllObjectProperties,
-  templateElement,
-  isUndefinedLiteral,
-  Primitive
-};
-
-// === TRANSFORMATION ===
-
-function traverseAST(code: Code, opts: TraverseOptions): t.File {
-  const ast: t.File = recast.parse(code, {
-    parser: {
-      parse: (source: Code) =>
-        parse(source, {
-          sourceType: "module",
-          allowImportExportEverywhere: true,
-          allowReturnOutsideFunction: true,
-          startLine: 1,
-
-          // Tokens are necessary for Recast to do its magic ✨
-          tokens: true,
-
-          plugins: [
-            "asyncGenerators",
-            "bigInt",
-            "classPrivateMethods",
-            "classPrivateProperties",
-            "classProperties",
-            "decorators-legacy",
-            "doExpressions",
-            "dynamicImport",
-            "exportDefaultFrom",
-            "exportNamespaceFrom",
-            "functionBind",
-            "functionSent",
-            "importMeta",
-            "nullishCoalescingOperator",
-            "numericSeparator",
-            "objectRestSpread",
-            "optionalCatchBinding",
-            "optionalChaining",
-            ["pipelineOperator", { proposal: "minimal" }],
-            "throwExpressions",
-            "jsx",
-            "typescript"
-          ]
-        })
-    }
-  });
-
-  traverse(ast, opts);
-
-  return ast;
-}
-
-function transform(code: Code, options: TraverseOptions): Transformed {
-  const ast = traverseAST(code, options);
-  const newCode = recast.print(ast).code;
-
-  return {
-    code: newCode,
-    hasCodeChanged: newCode !== code
-  };
-}
-
-interface Transformed {
-  code: Code;
-  hasCodeChanged: boolean;
-}
-
-// === SELECTION ===
-
-interface ASTSelection {
-  start: ASTPosition;
-  end: ASTPosition;
-}
-
-interface ASTPosition {
-  line: number;
-  column: number;
-}
-
-type SelectablePath = NodePath<SelectableNode>;
-type SelectableNode = Selectable<t.Node>;
-type SelectableObjectProperty = Selectable<t.ObjectProperty>;
-type SelectableIdentifier = Selectable<t.Identifier>;
-type SelectableVariableDeclarator = Selectable<t.VariableDeclarator>;
-type Selectable<T> = T & { loc: t.SourceLocation };
-
-function isSelectableNode(node: t.Node | null): node is SelectableNode {
-  return !!node && !!node.loc;
-}
-
-function isSelectableIdentifier(node: t.Node): node is SelectableIdentifier {
-  return t.isIdentifier(node) && isSelectableNode(node);
-}
-
-function isSelectableVariableDeclarator(
-  declaration: t.VariableDeclarator
-): declaration is SelectableVariableDeclarator {
-  return !!declaration.loc;
-}
-
-// === AST DOMAIN ===
-
-function isArrayExpressionElement(
-  node: t.Node | null
-): node is null | t.Expression | t.SpreadElement {
-  return node === null || t.isExpression(node) || t.isSpreadElement(node);
-}
-
-function areAllObjectProperties(
-  nodes: (t.Node | null)[]
-): nodes is t.ObjectProperty[] {
-  return nodes.every(node => t.isObjectProperty(node));
-}
-
-/**
- * Override babel `templateElement()` because it exposes
- * unnecessary implementation details and it's not type-safe.
- */
-function templateElement(value: string | number | boolean): t.TemplateElement {
-  return t.templateElement({
-    raw: value,
-    cooked: value
-  });
-}
-
-function isUndefinedLiteral(
-  node: object | null | undefined,
-  opts?: object | null
-): node is t.Identifier {
-  return t.isIdentifier(node, opts) && node.name === "undefined";
-}
-
-type Primitive =
-  | t.StringLiteral
-  | t.NumberLiteral
-  | t.BooleanLiteral
-  | t.BigIntLiteral;
