+++ src/ast/selection.ts
@@ -1,0 +1,47 @@
+import { NodePath } from "@babel/traverse";
+import * as t from "@babel/types";
+
+export {
+  ASTSelection,
+  ASTPosition,
+  SelectablePath,
+  SelectableNode,
+  SelectableObjectProperty,
+  SelectableIdentifier,
+  SelectableVariableDeclarator,
+  Selectable,
+  isSelectableNode,
+  isSelectableVariableDeclarator,
+  isSelectableIdentifier
+};
+
+interface ASTSelection {
+  start: ASTPosition;
+  end: ASTPosition;
+}
+
+interface ASTPosition {
+  line: number;
+  column: number;
+}
+
+type SelectablePath = NodePath<SelectableNode>;
+type SelectableNode = Selectable<t.Node>;
+type SelectableObjectProperty = Selectable<t.ObjectProperty>;
+type SelectableIdentifier = Selectable<t.Identifier>;
+type SelectableVariableDeclarator = Selectable<t.VariableDeclarator>;
+type Selectable<T> = T & { loc: t.SourceLocation };
+
+function isSelectableNode(node: t.Node | null): node is SelectableNode {
+  return !!node && !!node.loc;
+}
+
+function isSelectableIdentifier(node: t.Node): node is SelectableIdentifier {
+  return t.isIdentifier(node) && isSelectableNode(node);
+}
+
+function isSelectableVariableDeclarator(
+  declaration: t.VariableDeclarator
+): declaration is SelectableVariableDeclarator {
+  return !!declaration.loc;
+}
